requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-dotenv==1.1.1
google-genai==1.41.0
python-multipart==0.0.20
pillow==11.3.0
pydantic[email]==2.12.0
pydantic-settings==2.11.0
stripe==13.0.1
SQLAlchemy==2.0.44